Make script changes for cleanliness and testing

+++ scripts/step-5.sh
@@ -7,25 +7,35 @@ if [ -z "$GITHUB_USER" ]; then
     echo "Set GITHUB_USER"
     exit
 fi
-if [ -z "$CEREMONY_DATE" ]; then
-    CEREMONY_DATE=$(date '+%Y-%m-%d')
+if [ -z "$REPO" ]; then
+    REPO=$(pwd)/ceremony/$(date '+%Y-%m-%d')
+    echo "Using default REPO $REPO"
 fi
-export REPO=$(pwd)/ceremony/$CEREMONY_DATE
+
+if [ -z "$BRANCH" ]; then
+    export BRANCH=main
+else
+    echo "Using branch $BRANCH"
+fi
 
 # Dump the git state
 git status
 git remote -v
 
 git clean -d -f
-git checkout main
-git pull upstream main
+git checkout $BRANCH
+git pull upstream $BRANCH
 git status
 
 # Sign the root and targets
 ./tuf publish -repository $REPO
 # Clear and copy into the repository/
-rm -r repository/
-cp -r $REPO/repository/. repository/
+rm -r repository/ if [ -f $REPO/repository/1.root.json ]; then
+cp -r $REPO/repository/ repository/
+
+if [ -n "$NO_PUSH" ]; then
+    echo "Skipping push, exiting early..."
+fi
 
 git checkout -b publish
 git add ceremony/
@@ -33,4 +43,7 @@ git commit -s -a -m "Publishing for ${GITHUB_USER}!"
 git push -f origin publish
 
 # Open the browser
-open "https://github.com/${GITHUB_USER}/root-signing/pull/new/publish" || xdg-open "https://github.com/${GITHUB_USER}/root-signing/pull/new/publish"
+export GITHUB_URL=$(git remote -v | awk '/^upstream/{print $2}'| head -1 | sed -Ee 's#(git@|git://)#https://#' -e 's@com:@com/@' -e 's#\.git$##')
+export CHANGE_BRANCH=$(git symbolic-ref HEAD | cut -d"/" -f 3,4)
+export PR_URL=${GITHUB_URL}"/compare/${BRANCH}..."${CHANGE_BRANCH}"?expand=1"
+open "${PR_URL}" || xdg-open "${PR_URL}"
